First commit, version 8
I coded this years ago. I needed help to work with many simultaneous projects. I wanted to be able to go back to previous account numbers, or images, that I copied to the clipboard. Then the project became a swissknife.. to allow me zoom part of the screen, or to copy-paste part of the screen freezing it first. Finally I added an Evaluator (a calculator that calculates on the fly without having to press ENTER.. something quite prehistoric by now.. why we need to confirm!?). Another handy tool is a reminder to let you know when you worked too long without a break.