Source code version 8.6
Improvement in timing (big objects in clipboard will not be refresh as often to reduce CPU usage).  Selecting items from the Action popup (right click in history button) will not replace the current clipboard anymore (that allows opening a file and paste the current content). Improvement in the Evaluator screen and added RoundTo(n,d) function. Improvement in the Zoom screen (shift key to refresh, and alt to freeze). Other minor adjustments.